requirements.txt

flask==3.0.3
supabase-py==2.4.0
python-dotenv==1.0.1
